Make EXP430G2 Hal.c work with msp430-gcc 4.9.1
msp430-gcc 4.9.1 removed a few builtin functions and added some
new ones. Examples do not compile with the new gcc. This commit
updates Hal.c to work with "new" builtin functions, that are
in fact closer to the TI compiler's.

Updated code does not check for the version of the compiler, so
it stops working with older versions of msp430-gcc. It may be
worth to make changes conditional of the version of gcc, but I
do not have the old version to test, thus I did not do that.